# web/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCA API
  version: 1.0.0
  description: |-
    Bank statements in, approved double-entry journal entries out.

    ### Two conventions worth knowing before you read anything else

    **Money is always a whole number of paise**, in a field named `*_paise`. Every amount also carries a `*_display` twin, already formatted with Indian digit grouping (`Rs 6,03,490.57`). Use the integer for arithmetic and the string for rendering, and never parse the string or format the integer yourself -- a JSON number with a decimal point becomes a float in a browser, which is exactly the rounding error the backend exists to avoid.

    **Nothing is final until a senior CA approves it.** Uploading a statement and classifying its rows produces *suggestions*. Only `POST /clients/{id}/approvals/` writes to the ledger, only a senior CA or firm admin may call it, and what it writes cannot afterwards be edited or deleted -- corrections are new entries that reverse and replace.

    ### Slow work

    Uploading a statement returns **202 Accepted** with a job. Poll `/jobs/{id}/` or subscribe to `/jobs/{id}/events/` for server-sent events.

    ### Errors

...
paths:
...
  /api/v1/team/clients/{id}/team/{member_id}/:
    post:
      operationId: team_clients_team_create_2
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: member_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - team
      security:
      - cookieAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: team_clients_team_destroy_2
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: member_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - team
      security:
      - cookieAuth: []
      responses:
        '204':
          description: No response body
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
...